EPBDS-6776 Use properties as is without normalization. Support relative folders in install wizard.
--HG--
branch : 5.19.x

+++ STUDIO/org.openl.rules.webstudio/src/org/openl/rules/webstudio/web/install/InstallWizard.java
@@ -53,7 +53,7 @@ public class InstallWizard {
 
     @NotBlank
     private String workingDir;
-    private boolean newWorkingDir;
+    private boolean workingDirChanged;
     private boolean showErrorMessage = false;
 
     private String userMode = "single";
@@ -90,7 +90,7 @@ public class InstallWizard {
     public InstallWizard() {
         appConfig = new ConfigurationManager(true,
                 System.getProperty("webapp.root") + "/WEB-INF/conf/config.properties");
-        workingDir = appConfig.getPath("webstudio.home");
+        workingDir = appConfig.getStringProperty("webstudio.home");
 
         ServletContext servletContext = FacesUtils.getServletContext();
         dbUtils = new DBUtils(servletContext);
@@ -125,10 +125,8 @@ public class InstallWizard {
         // Go to next step
         ++step;
         if (step == 2) {
-            workingDir = ConfigurationManager.normalizePath(workingDir);
-
             // Get defaults from 'system.properties'
-            if (newWorkingDir || systemConfig == null) {
+            if (workingDirChanged || systemConfig == null) {
                 systemConfig = new ConfigurationManager(true,
                         workingDir + "/system-settings/system.properties",
                         System.getProperty("webapp.root") + "/WEB-INF/conf/system.properties");
@@ -231,7 +229,7 @@ public class InstallWizard {
             designRepositoryConfiguration.save();
 
             System.clearProperty("webstudio.home"); // Otherwise this property will not be saved to file.
-            appConfig.setPath("webstudio.home", workingDir);
+            appConfig.setProperty("webstudio.home", workingDir);
             appConfig.setProperty("webstudio.configured", true);
             appConfig.save();
             System.setProperty("webstudio.home", workingDir);
@@ -302,7 +300,7 @@ public class InstallWizard {
         File studioDir;
 
         if (StringUtils.isNotEmpty((String) value)) {
-            studioPath = ConfigurationManager.normalizePath((String) value);
+            studioPath = (String) value;
             studioDir = new File(studioPath);
 
             if (studioDir.exists()) {
@@ -324,7 +322,7 @@ public class InstallWizard {
                     throw new ValidatorException(FacesUtils.createErrorMessage("'" + studioPath + "' is not a folder"));
                 }
             } else {
-                File parentFolder = studioDir.getParentFile();
+                File parentFolder = studioDir.getAbsoluteFile().getParentFile();
                 File existingFolder = null;
 
                 while (parentFolder != null) {
@@ -376,16 +374,19 @@ public class InstallWizard {
      * @param studioFolder   folder were studio will be installed
      */
     private void deleteFolder(File existingFolder, File studioFolder) {
-
         if (!studioFolder.delete()) {
             log.warn("Can't delete the folder {}", studioFolder.getName());
+        }
+
+        if (existingFolder == null) {
+            return;
         }
 
         while (!studioFolder.getAbsolutePath().equalsIgnoreCase(existingFolder.getAbsolutePath())) {
             if (!studioFolder.delete()) {
                 log.warn("Can't delete the folder {}", studioFolder.getName());
             }
-            studioFolder = studioFolder.getParentFile();
+            studioFolder = studioFolder.getAbsoluteFile().getParentFile();
         }
     }
 
@@ -516,9 +517,8 @@ public class InstallWizard {
     }
 
     public void setWorkingDir(String workingDir) {
-        String normWorkingDir = ConfigurationManager.normalizePath(workingDir);
-        newWorkingDir = !normWorkingDir.equals(this.workingDir);
-        this.workingDir = normWorkingDir;
+        workingDirChanged = !workingDir.equals(this.workingDir);
+        this.workingDir = workingDir;
 
         // Other configurations depend on this property
         System.setProperty("webstudio.home", this.workingDir);
